continuous tasks added

+++ auth_api/src/db/task_repository.py
@@ -1,6 +1,8 @@
 from ..db.db import Database
 from sqlalchemy.orm import sessionmaker
 from ..entities.models import TaskDB, UserDB
+from sqlalchemy.exc import SQLAlchemyError
+from sqlalchemy import inspect
 
 class TaskRepository:
     def __init__(self):
@@ -14,7 +16,7 @@ class TaskRepository:
         try:
             session.add(task)
             session.commit()
-        except Exception as e:
+        except SQLAlchemyError as e:
             session.rollback()
             raise e
         finally:
@@ -34,16 +36,43 @@ class TaskRepository:
         finally:
             session.close()
 
-    def delete_task(self, id : int):
-        task = self.get_task_by_id(id)
+    def delete_task(self, id: int):
         session = self.db.get_session()
+        task = session.query(TaskDB).filter(TaskDB.id == id).first()
+        if not task:
+            raise ValueError(f"Task with id {id} not found")
         try:
             session.delete(task)
             session.commit()
             return task
-        except Exception as e:
+        except SQLAlchemyError as e:
             session.rollback()
             raise e
         finally:
             session.close()
 
+    def update_task(self, task_id: int, **kwargs):
+        session = self.db.get_session()
+        task = self.get_task_by_id(task_id)
+
+        if not task:
+            raise ValueError("Task not found")
+
+        try:
+            for key, value in kwargs.items():
+                if hasattr(task, key):  
+                    setattr(task, key, value)
+            
+            session.add(task)
+
+            state = inspect(task)
+            if state.modified:
+                session.commit()
+                session.refresh(task)  
+
+            return task
+        except SQLAlchemyError as e:
+            session.rollback()
+            raise e
+        finally:
+            session.close()
